+++ source/extensions/filters/network/redis_proxy/config.h
@@ -1,8 +1,6 @@
 #pragma once
 
 #include <string>
-#include <map>
-#include <list>
 
 #include "envoy/api/api.h"
 #include "envoy/config/core/v3/base.pb.h"